style: Updated header design for discover(servers) page
