+++ src/app/(public)/servers/detail/[id]/page.tsx
@@ -1,6 +1,9 @@
+"use client"
+
 import Image from "next/image"
 import Link from "next/link"
 import { CalendarDays, FacebookIcon, InstagramIcon, LaptopMinimal, Laugh, MessageSquareMore, Newspaper, PartyPopper, TwitchIcon, TwitterIcon, YoutubeIcon } from "lucide-react"
+import { placeholderImageUrl } from "@/data"
 import { PageLayout } from "@/components/main/page-layout"
 import ServerBannerBottom from "@/components/servers/servers-banner-bottom"
 import { Button } from "@/components/ui/button"
@@ -14,15 +17,14 @@ interface ServerDetailPageProps {
 const ServerDetailPage = ({
     params
 }: ServerDetailPageProps) => {
-    const placeholderImage = "/images/placeholder.svg"
     const serverInviteUrl = "/invite/server-invite-code"
 
     return (
         <PageLayout
             header={{
-                variant: "transparent",
                 container: "normal",
-                showNav: true,
+                position: "static",
+                navStyle: "basic",
             }}
         >
             <div className="relative">
@@ -31,7 +33,7 @@ const ServerDetailPage = ({
                         <div className="w-full h-[178px] lg:h-[320px] max-h-[320px] absolute top-0 bg-[#404eed]"></div>
                         <div className="mt-24 lg:mt-[120px] px-6 md:px-10 w-full max-w-[1260px] grid grid-cols-[repeat(4,1fr)] lg:grid-cols-[repeat(12,1fr)] gap-x-5 z-0">
                             <Image
-                                src={placeholderImage}
+                                src={placeholderImageUrl}
                                 alt={"Discovery splash banner for Server Name Discord server"}
                                 width={1180}
                                 height={400}
@@ -42,7 +44,7 @@ const ServerDetailPage = ({
                         <div className="z-[1]">
                             <div className="mt-[-52px] px-6 md:px-10 w-full max-w-[1260px] hidden lg:grid grid-cols-[repeat(4,1fr)] lg:grid-cols-[repeat(12,1fr)] gap-x-5">
                                 <Image
-                                    src={placeholderImage}
+                                    src={placeholderImageUrl}
                                     alt={"Discovery icon for Server Name Discord server"}
                                     width={104}
                                     height={104}
@@ -54,7 +56,7 @@ const ServerDetailPage = ({
                                 <div className="lg:ml-12 flex flex-col col-span-12 lg:col-span-9">
                                     <div className="flex flex-row">
                                         <Image
-                                            src={placeholderImage}
+                                            src={placeholderImageUrl}
                                             alt={"Discovery icon for Server Name Discord server"}
                                             width={48}
                                             height={48}
@@ -65,7 +67,7 @@ const ServerDetailPage = ({
                                             <div className="flex flex-row items-center">
                                                 <div className="mr-2 mb-1 lg:mb-2">
                                                     <div className="size-4 lg:size-6 relative flex justify-center items-center text-app-bg-brand-experiment">
-                                                        <svg className="size-full" aria-hidden="false" width="16" height="16" viewBox="0 0 16 16"><path fill="currentColor" fill-rule="evenodd" d="m16 7.6c0 .79-1.28 1.38-1.52 2.09s.44 2 0 2.59-1.84.35-2.46.8-.79 1.84-1.54 2.09-1.67-.8-2.47-.8-1.75 1-2.47.8-.92-1.64-1.54-2.09-2-.18-2.46-.8.23-1.84 0-2.59-1.54-1.3-1.54-2.09 1.28-1.38 1.52-2.09-.44-2 0-2.59 1.85-.35 2.48-.8.78-1.84 1.53-2.12 1.67.83 2.47.83 1.75-1 2.47-.8.91 1.64 1.53 2.09 2 .18 2.46.8-.23 1.84 0 2.59 1.54 1.3 1.54 2.09z"></path></svg>
+                                                        <svg className="size-full" aria-hidden="false" width="16" height="16" viewBox="0 0 16 16"><path fill="currentColor" fillRule="evenodd" d="m16 7.6c0 .79-1.28 1.38-1.52 2.09s.44 2 0 2.59-1.84.35-2.46.8-.79 1.84-1.54 2.09-1.67-.8-2.47-.8-1.75 1-2.47.8-.92-1.64-1.54-2.09-2-.18-2.46-.8.23-1.84 0-2.59-1.54-1.3-1.54-2.09 1.28-1.38 1.52-2.09-.44-2 0-2.59 1.85-.35 2.48-.8.78-1.84 1.53-2.12 1.67.83 2.47.83 1.75-1 2.47-.8.91 1.64 1.53 2.09 2 .18 2.46.8-.23 1.84 0 2.59 1.54 1.3 1.54 2.09z"></path></svg>
                                                         <div className="absolute inset-[-.05px,0,0.05px] flex justify-center items-center pointer-events-none">
                                                             <svg className="size-4 lg:size-6 text-app-white" aria-hidden="false" width="16" height="16" viewBox="0 0 16 16"><path d="M10.5906 6.39993L9.19223 7.29993C8.99246 7.39993 8.89258 7.39993 8.69281 7.29993C8.59293 7.19993 8.39317 7.09993 8.29328 6.99993C7.89375 6.89993 7.5941 6.99993 7.29445 7.19993L6.79504 7.49993L4.29797 9.19993C3.69867 9.49993 2.99949 9.39993 2.69984 8.79993C2.30031 8.29993 2.50008 7.59993 2.99949 7.19993L5.99598 5.19993C6.79504 4.69993 7.79387 4.49993 8.69281 4.69993C9.49188 4.89993 10.0912 5.29993 10.5906 5.89993C10.7904 6.09993 10.6905 6.29993 10.5906 6.39993Z" fill="currentColor"></path><path d="M13.4871 7.79985C13.4871 8.19985 13.2874 8.59985 12.9877 8.79985L9.89135 10.7999C9.29206 11.1999 8.69276 11.3999 7.99358 11.3999C7.69393 11.3999 7.49417 11.3999 7.19452 11.2999C6.39545 11.0999 5.79616 10.6999 5.29674 10.0999C5.19686 9.89985 5.29674 9.69985 5.39663 9.59985L6.79499 8.69985C6.89487 8.59985 7.09463 8.59985 7.19452 8.69985C7.39428 8.79985 7.59405 8.89985 7.69393 8.99985C8.09346 8.99985 8.39311 8.99985 8.69276 8.79985L9.39194 8.39985L11.3896 6.99985L11.6892 6.79985C12.1887 6.49985 12.9877 6.59985 13.2874 7.09985C13.4871 7.39985 13.4871 7.59985 13.4871 7.79985Z" fill="currentColor"></path></svg>
                                                         </div>
@@ -142,11 +144,13 @@ const ServerDetailPage = ({
                                             </div>
                                         </div>
                                     </div>
-                                    <div className="mb-6 lg:mb-[25px] pt-6 lg:pt-[25px] flex flex-col border-t-[1px] border-solid border-[rgba(6,6,7,.08)] [&_div.server-text-section]:mt-6 [&_div.server-text-section:first-child]:mt-0">
+                                    <div className="mb-6 lg:mb-[25px] pt-6 lg:pt-[25px] flex flex-col border-t-[1px] border-solid border-[rgba(6,6,7,.08)]
+                                    [&_div.server-text-section]:mt-6 [&_div.server-text-section:first-child]:mt-0">
                                         <div className="server-text-section">
                                             <h2 className="mb-2 text-app-header-primary font-display font-bold text-[20px] leading-6">About</h2>
                                             <div className="text-app-text-normal-2 text-[14px] lg:text-[16px] leading-[18px] lg:leading-5">
-                                                <div className="font-primary text-inherit select-text [&_p]:mt-2 [&_p:first-child]:mt-0">
+                                                <div className="font-primary text-inherit select-text
+                                                [&_p]:mt-2 [&_p:first-child]:mt-0">
                                                     <p>We are the largest community-run Server Name Discord. Join for news, chat, LFG, events &amp; more! We have channels for both Players and Creators.</p>
                                                     <p>Our server features Server Name inspired emoji and stickers, which you can use in other servers if you have Discord nitro.</p>
                                                     <p>Be sure to read the rules before chatting in the server!</p>
