+++ src/components/servers/server-info-card-item.tsx
@@ -1,5 +1,5 @@
 import { Suspense } from "react"
-import ServerInfoCardSkeleton from "@/components/skeletons/server-info-card-skeleton"
+import ServerInfoCardSkeleton from "@/components/skeleton/server-info-card-skeleton"
 import ServerInfoCard from "@/components/servers/server-info-card"
 
 const ServerInfoCardItem = () => {
