+++ src/components/skeleton/server-info-card-skeleton.tsx
